requirements.txt
jira
google-api-python-client
oauth2client
python-dateutil
pytz
